many changes for fucking sake

+++ app/models/Event.php
@@ -120,18 +120,26 @@
         }
 
         public function deleteEvent($id){
-
-            $this->db->query('DELETE FROM events where id= (:id)');
+            $this->db->query('SELECT * FROM events where id= (:id)');
             $this->db->bind(':id', $id);
-
-            if($this->db->execute()){
-                return true;
+            $row = $this->db->single();
+            if($this->db->rowCount() > 0 ){
+               $img = $row->image;
             }
             else{
                 return false;
             }
+            if(unlink(IMAGEROOT.$img)){
+                $this->db->query('DELETE FROM events where id= (:id)');
+                $this->db->bind(':id', $id);
 
+                if($this->db->execute()){
+                    return true;
+                }
+                else{
+                    return false;
+                }
+            }
+            
         }
-
-        
     }
